requirements.txt
Flask==2.3.3
pymongo==4.5.0
xlsxwriter==3.1.9
itsdangerous==2.1.2
requests==2.31.0
python-dotenv==1.0.0
